requirements.txt
# Core dependencies
llama-cpp-python>=0.3.0
pydantic>=2.0.0
pyyaml>=6.0

# Server (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Agent features (optional)
openai>=1.0.0
pyautogen>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.2.0
pytest-cov>=4.1.0
